fix: better Common abstraction
